refactor: extract solving submodule
